refactor: изменение логики хранения и использования моделей компетенций и технологий во избежание циклического импорта при построении связей
