+++ src/external/learning_analytics/data_formalization_submodule/urls.py
@@ -1,5 +1,13 @@
 from django.urls import path
 from src.external.learning_analytics.data_formalization_submodule.views import (
+    CompetencySendView,    
+    CompetencyGetView,     
+    CompetencyPutView,     
+    CompetencyDeleteView,  
+    TechnologySendView,
+    TechnologyGetView,
+    TechnologyPutView,
+    TechnologyDeleteView,
     SpecialityGetView,
     SpecialitySendView,
     SpecialityPutView,
@@ -19,6 +27,14 @@ from src.external.learning_analytics.data_formalization_submodule.views import (
 )
 
 urlpatterns = [
+    path('technologies/', TechnologyGetView.as_view(), name='technologies'),
+    path('technologies_send/', TechnologySendView.as_view(), name='technologies_send'),
+    path('technologies_put/<int:pk>/', TechnologyPutView.as_view(), name='technologies_put'),
+    path('technologies_delete/<int:pk>/', TechnologyDeleteView.as_view(), name='technologies_delete'),
+    path('competencies/', CompetencyGetView.as_view(), name='competencies'),
+    path('competencies_send/', CompetencySendView.as_view(), name='competencies_send'),
+    path('competencies_put/<int:pk>/', CompetencyPutView.as_view(), name='competencies_put'),
+    path('competencies_delete/<int:pk>/', CompetencyDeleteView.as_view(), name='competencies_delete'),
     path('specialities/', SpecialityGetView.as_view(), name='specialities'),
     path('send_specialitiy/', SpecialitySendView.as_view(), name='send_speciality'),
     path('speciality_put/<int:pk>/', SpecialityPutView.as_view(), name='speciality_put'),
